Complete Ch. 6
Not ꝥ it's in evidence anymore in the commit history (git rebase is fun(!)…)
but the fact ꝥ the URLs added to the Popscript aren't tested for correctness in Ch. 5
but then fail the Ch. 6 tests if they're off
(even by just omitting, as I accidentally did for most, the slash at the end)
is rather frustrating ngl.

Fortunately this won't be a public repo until I finish the task
(i.e. up to Ch. 10)
so `git push -f` will at least not throw any downstream users