Add Page‐adding view
Enough Magic is involved in all this Framework business
ꝥ I have no overview re whether it's possible for `Category.objects.get()` to return `None`.
If not, we could fold the ensuing conditional straight into the exception handler
but since I have no guarantee thereof it has to stay outside as per the example code

This whole ‘`category_name_slug`’ thing is really not very pleasant ☹︎
Words In Code is and will always be an antipattern
(needing symbolic rather than positional args is fine, but ꝥ's what single letters are for.
Command‐line flags (*not* GNU‐style long options) are the gold standard here.
(And I suppose trad‐Unix‐style abbreviated names are permissible too if your interface is *that* complex…
but three Full Unabbreviated WORDS? Not reasonable under any circumstances I mean come On))