Extend base Template
I object strongly to the substitution of a Hyphen–Minus
(a character which really should never appear in written human‐facing language at all —
we have both a Hyphen and a Minus sign in Unicode nowadays —
and whose use in Computer‐facing languages is excusable only by its being in ASCII)
for what should clearly be an m‐dash
in the <title> element — yet the tests check for it explicitly!

Fortunately that applies only to this chapter and the next,
so it's getting fixed once I hit Ch. 10