Refactor index Template
One big disadvantage of the url template tag is ꝥ in combination with ASCII‐only sluggifying,
the site is now very easy to break beyond anything but manual intervention:
the site doesn't validate attempted new categories for nonempty slugs,
so trying to create, say, the category ⟨!⟩ (or ⟨Ɂƿə⟩, or, or, or)
will cause the index page to fail at the reverse lookup for the empty slug,
and thereupon crash.
Previously it merely generated a (‐n albeit invald) URL ⟨/rango/category//⟩;
not great, but miles better than refusing to render anything.

Sort‐of almost acceptable for a toy application such as this,
but the usual lesson applies:
NEVER Trust User Input!