Add login Template
Another subtle(?) bug:
users are able to register with names of up to 150 chars (bytes? unicode code points? with or w/o normalisation?),
but the login form limits input to 50.
Allowing someone to pick a name but then not letting them use it
might make them want to avenge themselves.
Maybe by creating a category like ⟨×ↄ>>!⟩ and crashing your homepage…?

(That one time I became unable to log in to a system
because they changed the backend
and it couldn't handle double quotes ⟨"⟩ w/in a password comes to mind 😐︎)