Vary Footer links with Login status
The order of links varies from the example;
it rather matches how they were ordered in the example when the links were being introduced,
which tbf arguably also makes a bit more sense
(generic actions before auth‐related ones).
Some of the inconsistencies in this book seem a bit gratuitous ngl

But since the tests don't test the order I can get away with it.